YouTube AudioDownloader - python
I managed to make an application for download in mp3 youtube videos. Pytube downloads the file to mp4, so it is necessary to convert through some functions of the library os_sys. I defined thread functions for each download function. I had to do this due to the fact that tkinter puts everything in one thread only, and this causes the app to crash when we press the Download button. 
The code is ready to be compiled through auto-py-to-exe, as long as the pytube library is installed on your computer. The other libraries are already native to pyCharm.
You need an image .gif youtube.gif, or any other name you want, as long as you change the name in the code as well.